refactor: Extract duplicate ModelStub
